rename SB theme
